requirements.txt
Flask==2.2.3
azure-functions==1.14.0
pandas==1.5.3
scikit-learn==1.3.0
